+++ src/main/java/com/example/c11_examensarbete/services/ListService.java
@@ -5,6 +5,8 @@ import com.example.c11_examensarbete.dtos.ListDto;
 import com.example.c11_examensarbete.dtos.MangaDto;
 import com.example.c11_examensarbete.dtos.SavedMangaDto;
 import com.example.c11_examensarbete.entities.SavedManga;
+import com.example.c11_examensarbete.exceptionMappers.BadRequestExceptionMapper;
+import com.example.c11_examensarbete.exceptionMappers.ResourceNotFoundExceptionMapper;
 import com.example.c11_examensarbete.repositories.ListRepository;
 import com.example.c11_examensarbete.repositories.MangaRepository;
 import com.example.c11_examensarbete.repositories.SavedMangaRepository;
@@ -30,6 +32,9 @@ public class ListService {
     }
 
     public java.util.List<ListDto> getAllListsByUser(int id) {
+        if(id <= 0){
+            throw new BadRequestExceptionMapper("ID must be greater than 0.");
+        }
         return listRepository.findAll().stream()
                 .filter(list -> list.getUser().getId() == id)
                 .map(ListDto::fromList)
@@ -37,6 +42,9 @@ public class ListService {
     }
 
     public java.util.List<SavedMangaDto> getAllSavedMangasInList(int id) {
+        if (id <= 0){
+            throw new BadRequestExceptionMapper("ID must be greater than 0.");
+        }
         return listRepository.findById(id).stream()
                 .flatMap(list -> list.getSavedMangas().stream())
                 .map(SavedMangaDto::fromSavedManga)
@@ -44,6 +52,9 @@ public class ListService {
     }
 
     public java.util.List<MangaDto> getMangasFromSavedMangas(int listId) {
+        if(listId <= 0){
+            throw new BadRequestExceptionMapper("ID must be greater than 0.");
+        }
         java.util.List<Integer> mangaIds = listRepository.findById(listId).stream()
                 .flatMap(list -> list.getSavedMangas().stream())
                 .map(savedManga -> savedManga.getManga().getId())
@@ -55,10 +66,13 @@ public class ListService {
     }
 
     public int addList(ListDto listDto) {
+        if(listDto == null){
+            throw new BadRequestExceptionMapper("ListDto is null.");
+        }
         List list = new List();
         list.setListName(listDto.listName());
         list.setUser(userRepository.findById(listDto.userId())
-                .orElseThrow(() -> new IllegalArgumentException("User not found with id: " + listDto.userId())));
+                .orElseThrow(() -> new ResourceNotFoundExceptionMapper("User not found with id: " + listDto.userId())));
 
         listRepository.save(list);
 
@@ -67,24 +81,33 @@ public class ListService {
     }
 
     public int addSavedMangaToList(int listId, int savedMangaId) {
+        if(listId <= 0 || savedMangaId <= 0){
+            throw new BadRequestExceptionMapper("Both IDs must be greater than 0.");
+        }
         List list = listRepository.findById(listId)
-                .orElseThrow(() -> new IllegalArgumentException("List not found with id: " + listId));
+                .orElseThrow(() -> new ResourceNotFoundExceptionMapper("List not found with id: " + listId));
         SavedManga savedManga = savedMangaRepository.findByMangaId(savedMangaId).stream()
                 .filter(savedManga1 -> Objects.equals(savedManga1.getUser().getId(), list.getUser().getId()))
                 .findFirst()
-                .orElseThrow(() -> new IllegalArgumentException("SavedManga not found with id: " + savedMangaId));
+                .orElseThrow(() -> new ResourceNotFoundExceptionMapper("SavedManga not found with id: " + savedMangaId));
         list.getSavedMangas().add(savedManga);
         listRepository.save(list);
 
         return list.getId();
     }
     public void deleteList(int id) {
+        if (id <= 0){
+            throw new BadRequestExceptionMapper("ID must be greater than 0.");
+        }
         listRepository.deleteById(id);
     }
 
     public void deleteSavedMangaFromList(int listId, int mangaId) {
+        if(listId <= 0 || mangaId <= 0){
+            throw new BadRequestExceptionMapper("Both IDs must be greater than 0.");
+        }
         List list = listRepository.findById(listId)
-                .orElseThrow(() -> new IllegalArgumentException("List not found with id: " + listId));
+                .orElseThrow(() -> new ResourceNotFoundExceptionMapper("List not found with id: " + listId));
         list.getSavedMangas().removeIf(savedManga -> Objects.equals(savedManga.getId(), mangaId));
         listRepository.save(list);
     }
