+++ src/main/java/com/example/c11_examensarbete/services/SavedMangaService.java
@@ -1,17 +1,15 @@
 package com.example.c11_examensarbete.services;
 
-
-
 import com.example.c11_examensarbete.dtos.SavedMangaDto;
 import com.example.c11_examensarbete.entities.SavedManga;
+import com.example.c11_examensarbete.exceptionMappers.BadRequestExceptionMapper;
+import com.example.c11_examensarbete.exceptionMappers.ResourceNotFoundExceptionMapper;
 import com.example.c11_examensarbete.repositories.MangaRepository;
 import com.example.c11_examensarbete.repositories.SavedMangaRepository;
 import com.example.c11_examensarbete.repositories.UserRepository;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.NoSuchElementException;
-import java.util.Optional;
 
 @Service
 public class SavedMangaService {
@@ -27,6 +25,9 @@ public class SavedMangaService {
     }
 
     public List<SavedMangaDto> getUsersSavedMangas(int userid){
+        if(userid <= 0){
+            throw new BadRequestExceptionMapper("User id must be greater than 0");
+        }
         return savedMangaRepository.findAll().stream()
                 .filter(savedManga -> savedManga.getUser().getId() == userid)
                 .map(SavedMangaDto::fromSavedManga)
@@ -34,12 +35,15 @@ public class SavedMangaService {
     }
 
     public int saveManga(SavedMangaDto savedMangaDto) {
+        if(savedMangaDto == null) {
+            throw new BadRequestExceptionMapper("SavedMangaDto is null");
+        }
         System.out.println(savedMangaDto.mangaid());
         SavedManga savedManga = new SavedManga();
         savedManga.setManga(mangaRepository.findById(savedMangaDto.mangaid())
-                .orElseThrow(() -> new NoSuchElementException("Manga not found with id: " + savedMangaDto.mangaid())));
+                .orElseThrow(() -> new ResourceNotFoundExceptionMapper("Manga not found with id: " + savedMangaDto.mangaid() + ". Unable to save manga.")));
         savedManga.setUser(userRepository.findById(savedMangaDto.userid())
-                .orElseThrow(() -> new NoSuchElementException("User not found with id: " + savedMangaDto.userid())));
+                .orElseThrow(() -> new ResourceNotFoundExceptionMapper("User not found with id: " + savedMangaDto.userid() + ". Unable to associate user with saved manga.")));
 
         savedManga.setStatus(savedMangaDto.status());
         savedManga.setPersonalRating(savedMangaDto.score());
@@ -50,12 +54,15 @@ public class SavedMangaService {
     }
 
     public int editManga(SavedMangaDto savedMangaDto) {
+        if(savedMangaDto == null) {
+            throw new BadRequestExceptionMapper("SavedMangaDto is null");
+        }
         System.out.println("MangaId: " + savedMangaDto.mangaid());
         System.out.println("UserId: " + savedMangaDto.userid());
 
         List<SavedManga> mangaList = savedMangaRepository.findByMangaId(savedMangaDto.mangaid());
         if (mangaList.isEmpty()) {
-            throw new NoSuchElementException("No manga found with mangaId: " + savedMangaDto.mangaid());
+            throw new ResourceNotFoundExceptionMapper("No manga found with mangaId: " + savedMangaDto.mangaid());
         }
 
         mangaList.forEach(manga -> System.out.println("Found manga: " + manga));
@@ -63,7 +70,7 @@ public class SavedMangaService {
         SavedManga savedManga = mangaList.stream()
                 .filter(savedManga1 -> savedManga1.getUser().getId().equals(savedMangaDto.userid()))
                 .findFirst()
-                .orElseThrow(() -> new NoSuchElementException("Saved manga not found with id: " + savedMangaDto.mangaid() + " and user id: " + savedMangaDto.userid()));
+                .orElseThrow(() -> new ResourceNotFoundExceptionMapper("Saved manga not found with id: " + savedMangaDto.mangaid() + " and user id: " + savedMangaDto.userid()));
 
         savedManga.setStatus(savedMangaDto.status());
         savedManga.setPersonalRating(savedMangaDto.score());
@@ -76,10 +83,13 @@ public class SavedMangaService {
     }
 
     public void deleteSavedManga(int savedMangaId, int userId) {
+        if(savedMangaId <= 0 || userId <= 0){
+            throw new BadRequestExceptionMapper("SavedMangaId or UserId is less than 0");
+        }
         SavedManga savedManga = savedMangaRepository.findByMangaId(savedMangaId).stream()
                 .filter(savedManga1 -> savedManga1.getUser().getId() == userId)
                 .findFirst()
-                .orElseThrow(() -> new NoSuchElementException("Saved manga not found with id: " + savedMangaId));
+                .orElseThrow(() -> new ResourceNotFoundExceptionMapper("Saved manga not found with id: " + savedMangaId));
         savedMangaRepository.delete(savedManga);
     }
 }
